# backend/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Datify API",
    "description": "Self-hosted database branching platform - an open-source alternative to Neon",
    "contact": {
      "name": "Datify Contributors",
      "url": "https://github.com/datify/datify"
    },
    "license": {
      "name": "MIT",
      "url": "https://opensource.org/licenses/MIT"
    },
    "version": "0.1.0"
  },
  "servers": [
    {
      "url": "http://localhost:8080",
      "description": "Local development server"
    },
    {
      "url": "/",
      "description": "Current server"
    }
  ],
  "paths": {
...
    "/api/v1/databases/{id}": {
      "get": {
        "tags": [
          "Databases"
        ],
        "operationId": "get_database",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Database ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Database details",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - no access to database"
          },
          "404": {
            "description": "Database not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "put": {
        "tags": [
          "Databases"
        ],
        "operationId": "update_database",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Database ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateDatabaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Database updated successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - no access to database"
          },
          "404": {
            "description": "Database not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "delete": {
        "tags": [
          "Databases"
        ],
        "operationId": "delete_database",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Database ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "204": {
            "description": "Database deleted successfully"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - no access to database"
          },
          "404": {
            "description": "Database not found"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/databases/{id}/start": {
      "post": {
        "tags": [
          "Databases"
        ],
        "operationId": "start_database",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Database ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Database started successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - no access to database"
          },
          "404": {
            "description": "Database not found"
          },
          "409": {
            "description": "Database already running"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
    "/api/v1/databases/{id}/stop": {
      "post": {
        "tags": [
          "Databases"
        ],
        "operationId": "stop_database",
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "Database ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Database stopped successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseResponse"
                }
              }
            }
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - no access to database"
          },
          "404": {
            "description": "Database not found"
          },
          "409": {
            "description": "Database already stopped"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
    "/api/v1/projects/{project_id}/databases": {
      "get": {
        "tags": [
          "Databases"
        ],
        "operationId": "list_databases",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          },
          {
            "name": "limit",
            "in": "query",
            "description": "Number of items per page (default: 50)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "description": "Number of items to skip (default: 0)",
            "required": false,
            "schema": {
              "type": "integer",
              "format": "int64"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "List of databases with pagination"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not project owner"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      },
      "post": {
        "tags": [
          "Databases"
        ],
        "operationId": "create_database",
        "parameters": [
          {
            "name": "project_id",
            "in": "path",
            "description": "Project ID",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateDatabaseRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Database created successfully",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DatabaseResponse"
                }
              }
            }
          },
          "400": {
            "description": "Invalid request"
          },
          "401": {
            "description": "Unauthorized"
          },
          "403": {
            "description": "Forbidden - not project owner"
          }
        },
        "security": [
          {
            "bearer": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ConnectionInfo": {
        "type": "object",
        "required": [
          "host",
          "port",
          "username",
          "password",
          "database",
          "connection_string"
        ],
        "properties": {
          "connection_string": {
            "type": "string"
          },
          "database": {
            "type": "string"
          },
          "host": {
            "type": "string"
          },
          "password": {
            "type": "string"
          },
          "port": {
            "type": "integer",
            "format": "int32"
          },
          "username": {
            "type": "string"
          }
        }
      },
      "CreateDatabaseRequest": {
        "type": "object",
        "required": [
          "name"
        ],
        "properties": {
          "cpu_limit": {
            "type": "number",
            "format": "double"
          },
          "memory_limit_mb": {
            "type": "integer",
            "format": "int32"
          },
          "name": {
            "type": "string"
          },
          "postgres_version": {
            "type": "string",
            "example": "16"
          },
          "storage_limit_mb": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
...
      "DatabaseResponse": {
        "type": "object",
        "required": [
          "id",
          "project_id",
          "name",
          "postgres_version",
          "status",
          "resources",
          "created_at",
          "updated_at"
        ],
        "properties": {
          "connection": {
            "oneOf": [
              {
                "type": "null"
              },
              {
                "$ref": "#/components/schemas/ConnectionInfo"
              }
            ]
          },
          "created_at": {
            "type": "string"
          },
          "id": {
            "type": "string"
          },
          "name": {
            "type": "string"
          },
          "postgres_version": {
            "type": "string"
          },
          "project_id": {
            "type": "string"
          },
          "resources": {
            "$ref": "#/components/schemas/ResourceLimits"
          },
          "status": {
            "type": "string"
          },
          "updated_at": {
            "type": "string"
          }
        }
      },
...
      "ResourceLimits": {
        "type": "object",
        "required": [
          "cpu_limit",
          "memory_limit_mb",
          "storage_limit_mb"
        ],
        "properties": {
          "cpu_limit": {
            "type": "number",
            "format": "double"
          },
          "memory_limit_mb": {
            "type": "integer",
            "format": "int32"
          },
          "storage_limit_mb": {
            "type": "integer",
            "format": "int32"
          }
        }
      },
      "UpdateDatabaseRequest": {
        "type": "object",
        "properties": {
          "name": {
            "type": [
              "string",
              "null"
            ]
          }
        }
      },
...
    },
    "securitySchemes": {
      "bearer": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "JWT Bearer token authentication"
      }
    }
  },
...
}